# core/vm:  use rsh instead of lookup