This PR cleans up the code analysis a bit. There's this ugly lookup there, instead of a bitshift. 

For some reason, whenever I try to remove that lookup, like this PR does, there's a performance loss. Results from `go v1.17`, with `go test . -run - -bench JumpdestOp/\(PUSH1$\|STOP\) -count 10 ` 
```
name                        old time/op  new time/op  delta
JumpdestOpAnalysis/PUSH1-6  27.3ns ± 4%  31.0ns ± 0%  +13.41%  (p=0.000 n=10+8)
JumpdestOpAnalysis/STOP-6   22.0ns ± 8%  17.7ns ± 1%  -19.73%  (p=0.000 n=10+8)
```

So the `PUSH1` degrades `13%`, and the `STOP`, which _should not be affected at all_ improves by ~20%. 

:confused: 

Do other folks get similar numbers?