@holiman 2021-08-30
Tested again on top of the benchmark changes from @chfast -- but see the same thing.
```
name                        old time/op    new time/op    delta
JumpdestOpAnalysis/PUSH1-6    1.01ms ± 3%    1.17ms ± 3%  +16.49%  (p=0.000 n=10+10)
JumpdestOpAnalysis/STOP-6      758µs ± 2%     597µs ± 4%  -21.31%  (p=0.000 n=9+10)

name                        old speed      new speed      delta
JumpdestOpAnalysis/PUSH1-6  1.22GB/s ± 3%  1.05GB/s ± 3%  -14.16%  (p=0.000 n=10+10)
JumpdestOpAnalysis/STOP-6   1.61GB/s ± 3%  2.06GB/s ± 4%  +27.59%  (p=0.000 n=10+10)
```
For some reason, this change makes the worst-case slower.