@chfast 2021-08-30
I can benchmark this after #23499 is merged.