@chfast 2021-08-30
Note that `codeSegment()` does not use the lookup table.